+++ core/arch/arm/include/kernel/ldelf_loader.h
@@ -15,5 +15,9 @@ TEE_Result ldelf_init_with_ldelf(struct ts_session *sess,
 TEE_Result ldelf_dump_state(struct user_mode_ctx *uctx);
 TEE_Result ldelf_dump_ftrace(struct user_mode_ctx *uctx,
 			     void *buf, size_t *blen);
+TEE_Result ldelf_dlopen(struct user_mode_ctx *uctx, TEE_UUID *uuid,
+			uint32_t flags);
+TEE_Result ldelf_dlsym(struct user_mode_ctx *uctx, TEE_UUID *uuid,
+		       const char *sym, size_t maxlen, vaddr_t *val);
 
 #endif /* KERNEL_LDELF_LOADER_H */
